+++ vismop-vue/src/core/reactomeGraphs/reactomeOverviewNetwork/overviewTypes.d.ts
@@ -12,5 +12,21 @@ type fa2LayoutParams = {
   strongGravity: boolean;
   slowDown: number;
   barnesHutTheta: number;
-  clusterSizeScalingFactor: number
+  clusterSizeScalingFactor: number;
+};
+/**
+ * Type corresponding to the setting of one graph filter
+ */
+export type filterValues = {
+  limits: {
+    min: number;
+    max: number;
+  };
+  value: {
+    min: number;
+    max: number;
+  };
+  filterActive: boolean;
+  inside: boolean;
+  disable: boolean;
 };
