+++ vismop-vue/src/core/reactomeGraphs/reactomeTypes.d.ts
@@ -1,4 +1,5 @@
 // https://reactome.org/dev/diagram/pathway-diagram-specs for infos
+// types describing data structures in the reactome json files and detail diagrams
 import { upDatedPos } from '@/core/graphTypes';
 /*
 Basic Types
@@ -30,7 +31,6 @@ type reactomeEntry = {
   pathwayId: string;
   isCentral: boolean;
   clusterNum: number;
-  up: upDatedPos;
   rootId: string;
   initialPosX: number;
   initialPosY: number;
